+++ crates/tracel-core/Cargo.toml
@@ -12,7 +12,7 @@ categories.workspace = true
 
 [features]
 default = []
-station = ["burn-central-client/station", "tracel-experiment/station"]
+station = ["burn-central-client/station"]
 
 [dependencies]
 burn-central-client.workspace = true
@@ -25,3 +25,5 @@ directories = "6.0"
 toml = "1.0.0"
 crossbeam.workspace = true
 tracel-artifact.workspace = true
+tracing.workspace = true
+sha2.workspace = true
