+++ crates/tracel-core/src/experiment/cloud/artifacts.rs
@@ -1,26 +1,27 @@
-use burn_central_client::StationClient;
+use burn_central_client::Client;
 use tracel_artifact::bundle::FsBundle;
 
-use crate::{
+use tracel_experiment::{
     ArtifactKind, ExperimentId,
     reader::{ArtifactRef, ExperimentArtifactReader, ExperimentReaderError, LoadedArtifact},
-    remote::{
-        base::{ArtifactUploadError, ArtifactUploader},
-        station::{ExperimentArtifactClient, ExperimentPath},
-    },
 };
 
-pub struct StationArtifactReader {
-    client: StationClient,
+use crate::experiment::session::{ArtifactUploadError, ArtifactUploader};
+
+use super::{ExperimentArtifactClient, ExperimentPath};
+
+pub(crate) struct ConsoleArtifactReader {
+    client: Client,
+    exp_path: ExperimentPath,
 }
 
-impl StationArtifactReader {
-    pub fn new(client: StationClient) -> Self {
-        Self { client }
+impl ConsoleArtifactReader {
+    pub(crate) fn new(client: Client, exp_path: ExperimentPath) -> Self {
+        Self { client, exp_path }
     }
 }
 
-impl ExperimentArtifactReader for StationArtifactReader {
+impl ExperimentArtifactReader for ConsoleArtifactReader {
     fn load_artifact_raw(
         &self,
         experiment_id: ExperimentId,
@@ -30,7 +31,11 @@ impl ExperimentArtifactReader for StationArtifactReader {
             .parse::<i32>()
             .ok_or_else(|| ExperimentReaderError::new("Invalid experiment ID format"))?;
 
-        let experiment_path = ExperimentPath::new(num);
+        let experiment_path = ExperimentPath::new(
+            self.exp_path.owner_name().to_string(),
+            self.exp_path.project_name().to_string(),
+            num,
+        );
         let scope = ExperimentArtifactClient::new(self.client.clone(), experiment_path);
         let artifact = scope.fetch(name).map_err(|err| {
             ExperimentReaderError::with_source("Failed to resolve experiment artifact", err)
@@ -53,19 +58,19 @@ impl ExperimentArtifactReader for StationArtifactReader {
     }
 }
 
-pub struct StationArtifactUploader {
+pub(crate) struct ConsoleArtifactUploader {
     client: ExperimentArtifactClient,
 }
 
-impl StationArtifactUploader {
-    pub fn new(client: StationClient, exp_path: ExperimentPath) -> Self {
+impl ConsoleArtifactUploader {
+    pub(crate) fn new(client: Client, exp_path: ExperimentPath) -> Self {
         Self {
             client: ExperimentArtifactClient::new(client, exp_path),
         }
     }
 }
 
-impl ArtifactUploader for StationArtifactUploader {
+impl ArtifactUploader for ConsoleArtifactUploader {
     fn upload(
         &self,
         name: &str,
