+++ crates/tracel-experiment/Cargo.toml
@@ -11,10 +11,6 @@ repository.workspace = true
 keywords.workspace = true
 categories.workspace = true
 
-[features]
-default = ["station"]
-station = ["burn-central-client/station"]
-
 [dependencies]
 thiserror.workspace = true
 tracel-artifact.workspace = true
@@ -23,11 +19,6 @@ serde.workspace = true
 tracing.workspace = true
 tracing-subscriber = { workspace = true, features = ["env-filter"] }
 burn.workspace = true
-
-# Remote experiment dependencies
-burn-central-client = { workspace = true, features = ["central"] }
-crossbeam.workspace = true
-sha2.workspace = true
 
 [dev-dependencies]
 tokio = { version = "1", features = ["macros", "rt-multi-thread"] }
