--- crates/tracel-experiment/src/remote/mod.rs
@@ -1,6 +1,0 @@
-pub mod base;
-pub mod cloud;
-mod logs;
-mod socket;
-#[cfg(feature = "station")]
-pub mod station;
