+++ crates/tracel/Cargo.toml
@@ -13,7 +13,7 @@ rust-version.workspace = true
 
 [features]
 default = []
-station = ["tracel-experiment/station"]
+station = ["tracel-runtime/station"]
 
 [dependencies]
 tracel-macros.workspace = true
